Add Qemu builder.
Added a Qemu builder, although very hard-coded to my server. It
requires a VNC viewer to see the console. I could start one
on the laptop, connecting to vmhost01 (the server running the
Qemu VMs). I could see the console for a while, but then Qemu
or something decided that there was no console, and gave up.
I saw enough to be confident that the boot loader or whatever
was doing the same thing in Qemu as it does in Virtualbox.