Almost there
I added the post-processor to make the Vagrant box. I fixed the
authorization key so the ssh loging works now, and converts to
a secure key. I had to install the Virtaulbox guest addtions iso
to get the install to work. But it appears the guest addtions
aren't getting installed anyway, as the mount of /vagrant fails.