Turn on VirtualBox time sync.
Added the code in `template.json` to time sync when the VM resumes.
My assumption is that when the host goes to sleep (as it does when
it's a laptop), the VM suspends. On resume, it doesn't reset the
time in one jump, and so the time is no longer in sync with the
host.